requirements.txt
roboflow>=1.1.0
opencv-python>=4.8.0
pillow>=10.0.0
numpy>=1.24.0
scikit-image>=0.21.0
matplotlib>=3.7.0
tqdm>=4.65.0
pyyaml>=6.0
ultralytics>=8.1.0
